requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
language-check==1.1
leetpass==0.1.0
loremipsum==1.0.5
markovify==0.7.1
MarkupSafe==1.0
nltk==3.2.5
passwordstrength==0.1.1
PyYAML==3.12
requests==2.18.4
six==1.11.0
Unidecode==1.0.22
urllib3==1.22
Werkzeug==0.14.1
